+++ src/vacuum_device_accessory.ts
@@ -265,12 +265,12 @@ export class VacuumDeviceAccessory {
       this.log.info(`Room mapping found: ${JSON.stringify(roomMapping)}`);
       this.log.info(`Creating service areas from room mapping...`);
       return roomMapping.map(
-        ([roomName, roomId]): ServiceArea.Area => ({
+        ([roomId, roomName]): ServiceArea.Area => ({
           areaId: parseInt(roomId),
           mapId: null,
           areaInfo: {
             locationInfo: {
-              locationName: roomName,
+              locationName: `${roomName}`,
               floorNumber: null,
               areaType: null,
             },